#rb none

[CL 4835066 by Chris Gagnon in Dev-Editor branch]

+++ Engine/Source/Runtime/Core/Public/Misc/SecureHash.h
@@ -49,7 +49,7 @@ public:
 	 * @param input		input data
 	 * @param inputLen	length of the input data in bytes
 	 **/
-	void Update(const uint8* input, int32 inputLen);
+	void Update(const uint8* input, uint64 inputLen);
 
 	/**
 	 * MD5 finalization. Ends an MD5 message-digest operation, writing the
@@ -244,7 +244,7 @@ public:
 	void Reset();
 
 	// Update the hash value
-	void Update(const uint8 *data, uint32 len);
+	void Update(const uint8 *data, uint64 len);
 
 	// Update the hash value with string
 	void UpdateWithString(const TCHAR *data, uint32 len);
@@ -262,7 +262,7 @@ public:
 	 * @param DataSize Size of the Data block
 	 * @param OutHash Resulting hash value (20 byte buffer)
 	 */
-	static void HashBuffer(const void* Data, uint32 DataSize, uint8* OutHash);
+	static void HashBuffer(const void* Data, uint64 DataSize, uint8* OutHash);
 
 	/**
 	 * Generate the HMAC (Hash-based Message Authentication Code) for a block of data.
@@ -274,7 +274,7 @@ public:
 	 * @param DataSize	Size of the Data block
 	 * @param OutHash	Resulting hash value (20 byte buffer)
 	 */
-	static void HMACBuffer(const void* Key, uint32 KeySize, const void* Data, uint32 DataSize, uint8* OutHash);
+	static void HMACBuffer(const void* Key, uint32 KeySize, const void* Data, uint64 DataSize, uint8* OutHash);
 
 	/**
 	 * Shared hashes.sha reading code (each platform gets a buffer to the data,
